apps/backend/pyproject.toml
[project]
name = "farmcode"
version = "0.1.0"
description = "AI agent orchestration platform for complete SDLC management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "GitPython>=3.1",
    "aiofiles>=24.0",
    "mcp>=1.0.0",  # Model Context Protocol SDK
    "fastapi>=0.115.0",  # For MCP server HTTP transport
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.8",
    "types-PyYAML",
]

[project.scripts]
farmcode = "farmcode.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["farmcode"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "DTZ", "T20", "RET", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
